Revert: [-] Removed relations between Categories and Products

+++ app/Modules/Categories/Resources/views/categories.blade.php
@@ -25,13 +25,20 @@
             <a href='{!! route('categories.edit', ['id' => $category->id]); !!}' class="btn btn-primary">
                 <i class="fa fa-pencil"></i>
             </a>
-
+            @if(CategoriesHelper::canDelete($category))
                 {!! Form::open(['route' => ['categories.delete', $category->id], 'method' => 'delete','class' => 'inline']) !!}
                 <a href="#" class="btn btn-danger" onclick="deleteCategory(this)" data-toggle="tooltip"
                    title="Delete category">
                     <i class="fa fa-trash"></i>
                 </a>
                 {!! Form::close() !!}
+            @else
+                <span data-toggle="tooltip" title="Cannot delete category with products">
+                    <a href='#' class="btn btn-danger disabled">
+                        <i class="fa fa-trash"></i>
+                    </a>
+                </span>
+            @endif
 
         </div>
